Delete Test is working great.
Test is now using CheckTreeBlackHeight method. It counts the black height of the tree and checks at every node if the height is the same in both children.
It also checks if the parent of the node is actually what it should be.

Rotate is fixed and now it also updates the parent field in the new brother node.
There are two rotates. The new one does not get the node argument.
The node can be null (In delete case 3) so we use there a rotate method that does not use the node itself.

Because children of the brother could be null (and it's ok, its the same case) there is a new method that returns the color of a node. if the node is null we consider it to be black.

Also, in case 2, if parent was red, we should paint it black and stop the balance.